rename: CodeTableViewCell -> LabelTableViewCell
